validações de usuário e chamado

+++ src/main/java/com/soulcode/hercall/services/ChamadoService.java
@@ -1,8 +1,11 @@
 package com.soulcode.hercall.services;
 
 import com.soulcode.hercall.dtos.ChamadoDto;
+import com.soulcode.hercall.dtos.SetorDto;
 import com.soulcode.hercall.models.Chamado;
+import com.soulcode.hercall.models.Setor;
 import com.soulcode.hercall.repositories.ChamadoRepository;
+import com.soulcode.hercall.shared.ApiResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,41 +18,80 @@ public class ChamadoService {
 
     @Autowired
     ChamadoRepository chamadoRepository;
-    public ChamadoDto save(ChamadoDto dto) {
-        Chamado chamado = ChamadoDto.convert(dto);
-        chamado = this.chamadoRepository.save(chamado);
-        return  new ChamadoDto(chamado);
-    }
 
-    public List<ChamadoDto> findAll() {
+    public ApiResponse<ChamadoDto> save(ChamadoDto dto) {
+        try {
+            Chamado chamado = ChamadoDto.convert(dto);
+            chamado = this.chamadoRepository.save(chamado);
 
-        List<Chamado>chamado = this.chamadoRepository.findAll();
-        return chamado.stream().map(ChamadoDto::new).collect(Collectors.toList());
-    }
-
-    public ChamadoDto findById(Long id) {
-
-        Optional<Chamado> chamado = this.chamadoRepository.findById(id);
-        if(chamado.isEmpty()){
-            throw new RuntimeException("Cliente não encontrado");
-        }else{
-            return new ChamadoDto(chamado.get());
+            return new ApiResponse<>(201, "Chamado cadastrado com sucesso!", new ChamadoDto(chamado));
+        } catch (Exception e) {
+            return new ApiResponse<>(500, e.getMessage(), null);
         }
     }
 
-    public ChamadoDto updateById(Long id, ChamadoDto dto) {
-
-        this.findById(id);
-        Chamado chamado = ChamadoDto.convert(dto);
-        chamado.setId(id);
-        this.chamadoRepository.save(chamado);
-        return new ChamadoDto(chamado);
+    public ApiResponse<List<ChamadoDto>> findAll() {
+        try {
+            List<Chamado> chamados = this.chamadoRepository.findAll();
+            return new ApiResponse<>(200, "Listagem de chamados realizada com sucesso!",
+                    chamados.stream().map(ChamadoDto::new).collect(Collectors.toList()));
+        } catch (Exception e) {
+            return new ApiResponse<>(500, e.getMessage(), null);
+        }
     }
 
-    public String deleteById(Long id) {
+    public ApiResponse<ChamadoDto> findById(Long id) {
+        try {
+            Optional<Chamado> resultado = this.chamadoRepository.findById(id);
 
-        ChamadoDto dto = findById(id);
-        this.chamadoRepository.deleteById(id);
-        return ( " O chamado " + dto.getId() + " foi excluído");
+            return resultado.map(chamado -> new ApiResponse<>(200, "Detalhamento de chamados realizado com sucesso!",
+                    new ChamadoDto(chamado))).orElseGet(() -> new ApiResponse<>(204, "Chamado não encontrado!", null));
+        } catch (Exception e) {
+            return new ApiResponse<>(500, e.getMessage(), null);
+        }
     }
+
+    public ApiResponse<ChamadoDto> updateById(Long id, ChamadoDto dto) {
+        try {
+            ApiResponse<ChamadoDto> existeChamado = this.findById(id);
+
+            if (existeChamado.getStatus() != 200) {
+                return new ApiResponse<>(404, "Não é possível editar, pois chamado não foi encontrado por ID!", null);
+            }
+
+            if(existeChamado.getData().getStatus().getNome().equalsIgnoreCase("finalizado")) {
+                return new ApiResponse<>(400, "Não é possível editar, pois chamado já foi finalizado!", null);
+            }
+
+            Chamado chamado = ChamadoDto.convert(dto);
+            chamado.setId(id);
+
+            chamado = this.chamadoRepository.save(chamado);
+
+            return new ApiResponse<>(200, "Chamado editado com sucesso!", new ChamadoDto(chamado));
+        } catch (Exception e) {
+            return new ApiResponse<>(500, e.getMessage(), null);
+        }
+    }
+
+    public ApiResponse<ChamadoDto> deleteById(Long id) {
+        try {
+            ApiResponse<ChamadoDto> existeChamado = this.findById(id);
+
+            if (existeChamado.getStatus() != 200) {
+                return new ApiResponse<>(404, "Não foi possível excluir, pois chamado não foi encontrado por ID!", null);
+            }
+
+            if (chamadoRepository.existChamadoComReponsavel(id)) {
+                return new ApiResponse<>(400, "Não é possível excluir, pois um técnico está atendendo esse chamado!", null);
+            }
+
+            this.chamadoRepository.deleteById(id);
+
+            return new ApiResponse<>(200, "Chamado excluído com sucesso!", existeChamado.getData());
+        } catch (Exception e) {
+            return new ApiResponse<>(500, e.getMessage(), null);
+        }
+    }
+
 }
